bears/general: Add InvalidLinkBear
Check for broken links

Search for every link in a file, and send a GET
request to each of them to check if they are valid.

If any of them do not send a HTTP status 200 back, then:

- if the status code is 30x:
    the redirect url is fetched and a diff is shown to update the
    url.
- if the status code is 50x or 40x:
    The message "Broken link - unable to connect to this url."
    along with the status code is displayed.
- if the GET request fails:
    The message "Broken link - unable to connect to this url."
    is displayed.

InvalidLinkBear also checks for internet connection, as a
prerequisite.

Fixes https://github.com/coala-analyzer/coala-bears/issues/29