bears/r: Add R support using RLintBear
Create the RLintBear using the tool lintr using the Lint class.

We also add popular R scripts in the gitignore that are created
for settings.

We need to set LINTR_COMMENT_BOT to false **BEFORE** installing
lintr, otherwise the lintr-bot is actuvated, and this tries to
comment on github automitically (when it detects we run on CI).
This fails and causes errors for us.

Fixes https://github.com/coala-analyzer/coala-bears/issues/41