Bears: Add IndentationBear
It is a generic indent bear, which looks for a start and end
indent eg: { : }  if they are specified, and it
also check for false indents in strings and comments
eg: "{"  where { is an
indent start.
It also indents functions recursively  for eg:

func(param1,
     param2)

and

func(param1,
     param2(param3,
	    param4)
     param5
)

and also does some other minor but cool indenty things.

it does not however support hanging indents and indents based on
keywords yet for eg:

if(something)
    does not get indented

changes to:

if(something)
does not get indented

Fixes https://github.com/coala-analyzer/coala-bears/issues/40