LocalBearTestHelper: Force lists and tuples
Force the valid_files ad invalid_files to be tuples/lists.
Also force the lines of a file to be tuple/list and do not
auto convert a string to a list assuming it's a file with
a single line. This causes confusion when using
`verify_local_bear()` as the whole tuple of 1 element needs
extra comma thing is confusing.

Fixes https://github.com/coala-analyzer/coala-bears/issues/99