requirements.txt: Fix versions of packages
So we don't get dependency problems, this commit fixes the versions
again. Updating a version means it needs to be tested, providing `>=`
for packages doesn't guarantee a dependency works everytime.

If someone still wants to use a newer version of a dependency, he can
install it manually with a higher version, hoping that coala-bears or
its tests still work.