bears: Remove CheckPrerequisites
The Lint class has been updated with a newer prerequisite checking
function that checks for two new things:

prerequisite_command : this command, if exists, should return an
exitcode of 0.
prerequisite_fail_msg : message to display if exitcode != 0

This was introduced in:
https://github.com/coala-analyzer/coala/commit/188a942

This renders the CheckPrerequisites module invalid. This commit
removes this redundant class and also refactors bears that used
this module.

Fixes https://github.com/coala-analyzer/coala-bears/issues/146