PerlCriticBear: Replace inline config with file
PerlCriticBear currently gives an exception when ran because the
`--config` argument does not exist. This should be `--profile`
instead.

Also, due to a known bug in temporary file creation, the inline
configuration file in PerCriticBearTest is moved to a directory
testfiles/ and the configuration file is placed there.

Fixes https://github.com/coala-analyzer/coala-bears/issues/109