ClangCodeCloneDetectionBear: Add switch support
Add support for switch/case statements in ClangCodeCloneDetectionBear
as  it does not currenly support it.This would count the condition with
the Is condition CC and the bodies with the In condition CCs so nested
things are recognized right. This includes test cases.

Fixes https://github.com/coala-analyzer/coala-bears/issues/39