xml: Rename package to xml2
As we put each bear-directory and their subdirectories into the Python
path, our `xml` package shadows the built-in. This makes problems inside
Windows (making the default bears practically unusable).

coala encourages to not import each subdirectory so that built-ins are
not shadowed accidentally, though this is difficult. This is a hotfix so
Windows users can do something with coala again.