requirements: Remove setuptools
Pip automatically tells if setuptools is not installed when doing a
source install, and on wheel installs the setuptools is not needed as
the setup.py file doesn't exist anymore.
Also, coala will normally be installed using wheels in pip 6.x+ because
we upload wheels. Hence, we don't need setuptools.

The older versions of pip are anyway not being supported because we use
the ~= syntax. And the older versions of pip will just throw an import
error at the setuptools.

Hence, the setuptools in requirements.txt has no purpose other than
forcing the user to upgrade the setuptools above the lower version we
specify.

Fixes https://github.com/coala-analyzer/coala-bears/issues/751