ESLintBear: Handle corner case if eslint fails
If eslint fails to run, for example if it gets an invalid config file it
does not know how to handle, earlier the bear would throw a JSON
Decoding error because it wouldn't know how to parse "" (an empty
string). Now this case is handled, and it gracefull doesn't try to
create results in such cases.

Fixes https://github.com/coala-analyzer/coala-bears/issues/727