requirements: Use compatibility operator
It's better. Really. Google it. Read the PEP. It solves problems.

All requirements need at least a minor for this to work so versions that
aren't precisely pinned are pinned in this commit as well to their
latest version.

appdirs was not correctly pinned and is already in the requirements of
coala, however not at all actively used by coala-bears so it has been
removed.