AnnotationBear: Allow missing language setting
A missing language setting is not any different from
a language name which has a Language definition that
does not contain the necessary information to correctly
parse the file.

Use a special 'UnknownLanguage' in this scenario.

Also upgrades AnnotationBear from LanguageDefinition
to the new Language.

Fixes https://github.com/coala/coala-bears/issues/1012
Closes https://github.com/coala/coala-bears/issues/1095