Remove libclang-py3 from requirements.txt
This dependency is already managed by the coala package.
Due to continuous integration coala and coala-bears installing the
latest master of the other package, it can not be modified if it
appears in both otherwise a pkg_resources.ContextualVersionConflict
is raised when it is used.

Fixes https://github.com/coala/coala-bears/issues/1110