bear-requirements.txt: Use PipRequirement in bears
Remove all bear requirements from requirements.txt,
and use .ci/generate_bear_requirements.py to create
and verify that the bear-requirements is identical to
the PipRequirement found in the bears.

Update the PipRequirement in the bears to match the
versions currently being installed, and also remove
the PipRequirement for libclang-py, as this
dependency is provided by coala.

Updates setup.py to add an extra 'alldeps' that is be used
by CI to transition to explicitly requesting additional
dependencies so 100% coverage and `coala --ci` work correctly.

Uses pip to install wheels, as it supports extras while
setup.py would require a non-standardised command line option
to be added.

Closes https://github.com/coala/coala-bears/issues/1000