RadonBear: Pin to version 1.4.0
radon 1.4.2 added a dependency on flake8-polyfill, which depends
on flake8 without any version specification, and flake8 versions
have various version dependencies on pyflakes and pycodestyle.
As a result, it is not possible to know what versions of pyflakes
and pycodestyle that will be required to be compatible with
whatever flake8 version happens to be installed.

Fixes https://github.com/coala/coala-bears/issues/1228