yamllint: Disable `document-start` by default
Some projects use documents starts (`---`) in YAML documents (for
example Ansible, OpenStack, yamllint), others don't. Since these markers
are required when declaring multiple documents in a single .yaml file
(see the spec [1]), it is a bad idea to forbid them.

This commit disables the `document-start` rule by default, instead of
forcing / forbidding the use of these markers.

[1]: http://yaml.org/spec/1.2/spec.html#id2800132

Closes #1417
Related to #923 and #965