bears/__init__.py: Check version against coala
Print a warning if the installed version of package
coala is not same with coala-bears. Only check 
MAJOR.MINOR version.

Note that it's okay for the package bears to have 
higher version than coala. The patch level of each 
package runs independently, as those patches 
should be bugs only within that package, and 
released very quickly, and thus lots of bugs in 
a minor release of coalib could mean 
its patch level increments more quickly than 
the bears patch level.

Closes https://github.com/coala/coala/issues/3042