MarkdownBear: Set max_line_length default to None
The introduction of max_line_length with a default of 80
was a breaking change as previously line length was not
constrained.

This constraint required the use of remark-lint, which
caused MarkdownBear to remove all content from files
if remark-lint could not be found in NODE_PATH.

The default is now again to not check line length,
and not use remark-lint unless the .coafile explicitly
requests a max_line_length.

Fixes https://github.com/coala/coala-bears/issues/1581