Add FutureBear for running futurize on code
Closes https://github.com/coala/coala-bears/issues/1606

This bear will run futurize on code and present the user with
options to replace code with the equivalent Python2/3 compatible
statements.

NOTE: This does not working properly right now due to a bug
in coalib! The unified-diff output type is not preserving
new line characters in the corrected output.