RuboCopBear: Fix CI breakage for version > 0.47.1
https://github.com/bbatsov/rubocop/pull/4226 indicates that `--stdin`
takes filename as an argument which was incorporated. A part of the
fixing was to improve the tests that involved the bear to be run on
files with the `.rb`(ruby) extension because initially the bear was
run on the generated temporary files that was the cause of error.

Fixes https://github.com/coala/coala-bears/issues/1548