CPPCleanBear.py: Added include_path for header files
The feature for allowing include_path was added so the user can include paths for header files for cppclean. Only one include_path can be specified in the .coafile with this commit and the ability to add more include_path has not been added yet.

References https://github.com/coala/coala-bears/issues/2026