CPPCleanBear.py: Add include_paths setting
A new setting is added for the CPPCleanBear for allowing paths for
header files to be added in the .coafile. 'include_paths' is a
list of paths(strings) for header files which have been included in
the cpp file on which we desire to use cppclean. Tests,
'CPPCleanBearValidIncludePathTest' and
'CPPCleanBearInvalidIncludeSinglePathTest', are added.

Closes https://github.com/coala/coala-bears/issues/2026