JSONFormatBear.py: Try to fix float format
I made PR as requested in the issue. Some hints will be appreciated.
Tries to fix float representation by wrapping it in objects and keeping
them as strings. Then reads the string from this object and serializes
it in json file. However, float saved as string and each attempt
convert it to float results in missing the initial format of that
float.

Fixes https://github.com/coala/coala-bears/issues/2062