setup.cfg: Remove pytest config 'python_functions'
setup.cfg currently contains:

python_functions = *_test

This has the effect of allowing functions like `test(..)` to
exist and not be considered test functions.  Two sets have appeared.
When pytest defaults are used, these two sets of functions called
`test(..)` result in pytest loading failures as they are not valid
pytest test functions, but are helper functions for unittests.
The arguments on these functions should be fixture names, and they
are not.  This confuses people who know pytest naming conventions.