ESLintBearTest: Test default ignored file message
When eslint notices the source file is its own configuration
file, it emits a warning at the file level and without a keyId.
This test ensures that behaviour doesnt result in an unhandled
exception with the bear, and captures the output so it can
be handled correctly in a subsequent task.

Related to https://github.com/coala/coala-bears/issues/2533
Related to https://github.com/coala/coala-bears/issues/2607