bears/: Update required Python version to 3.4.4
coala updated its minimum version to 3.4.4 a
while ago and bears was not also updated.

Adds tests to cover assert_supported_version,
so that the pragma no cover can be removed
and setup.py does not need to be run by coverage.

Related to https://github.com/coala/coala/issues/4350