.travis.yml: Remove version numbers from PATH line
Create and use variables for the versions, so that
the PATH line does not need to be modified so often,
and the merge conflicts that causes.

Use ~/.local/ where possible for installed software,
so that scripts can be reused.

Related to https://github.com/coala/coala-bears/pull/1429
Related to https://github.com/coala/coala-bears/issues/2638