TSLintBearTest.py: Remove useless test
tslint 5 requires a configuration file, which breaks test
TSLintBearOtherOptionsTest and other tests.
However, this test supposedly provides coverage for 'rules_dir',
but it is provides '/' as the rules directory, which is invalid,
and does not confirm this functionality is actually working.

Replace with pragma no cover temporarily, until proper tests are
written.

Related to https://github.com/coala/coala-bears/issues/1072