TextLintBearTest: Remove unnecessary result
The test data for test_bad_alex_no_dead_link was causing
a error about gender, which was unnecessary in this test.
The underlying error has changed slightly in newer versions
of the linter, so removing the test case allows the bear
to pass tests on a wider array of linter versions.

Closes https://github.com/coala/coala-bears/issues/2777