CI: Fix Pycodestyle Warning W605
W605 is a warning for invalid escape sequences.

In python 3.6, Unknown escapes consisting of '\'
and an ASCII letter in regular expressions will
cause an warning.In the upcoming python versions
this would be treated as an error.
Owing to it pycodestyle have added this check for
future Python versions.

The alternative way to escape special characters
now is to append a extra `\` infront of the
sequences. Escape the characters by double backticks.

Fixes https://github.com/coala/coala-bears/issues/2760