PySafetyBear.py: Add proxy positional argument
The newer Safety package(v1.8.5) now adds the
functionality to support proxies which the Requests
package of Python uses.

The above functionality by Safety has been provided
by adding an extra positional argument called `proxy`
in the `saftey.check()` function. This caused the
TypeError in the previous version of Safety,due to
mismatch of number of positional argument

Closes https://github.com/coala/coala-bears/issues/2857