tox.ini: Create
This replaces Travis CI 'single image' with one VM for each set
of bears with similar dependencies, using tox to select the sets.

And it replaces AppVeyor CI to use FudgeCI to be a managable
'single image' VM that includes most of the dependencies needed
by bears, and uses tox to select the sets of bears.

The tox setup allows being used with or without --sitepackages,
and the AppVeyor CI does use --sitepackages and other tricks
to allow .ps1 finer control over the process of installing python
dependencies.

This uses tox-backticks to allow multiple selectors to be mapped
to multiple tests that pytest will run.

It uses pytest-cov-threshold to allow custom thresholds for each
source file, allowing each job to deselect the source files that
need coverage depending on the selected tests for the job.